extract sliver observer controller
